QueryBuilderの分割、READMEちょっと整備、テスト修正
