+++ delete_query_builder.go
@@ -1,0 +1,5 @@
+package query_builder
+
+type DeleteQueryBuilder struct {
+	*queryBuilder
+}
